+++ packages/create-sitecore-jss/src/templates/nextjs-sxa/src/pages/500.tsx
@@ -44,7 +44,7 @@ const Custom500 = (props: SitecorePageProps): JSX.Element => {
 };
 
 export const getStaticProps: GetStaticProps = async (context) => {
-  const site = siteResolver.getByName(config.siteName);
+  const site = siteResolver.getByName(config.sitecoreSiteName);
   const errorPagesService = new GraphQLErrorPagesService({
     clientFactory,
     siteName: site.name,