+++ packages/create-sitecore-jss/src/templates/node-headless-ssr-experience-edge/src/config.ts
@@ -1,13 +1,14 @@
 import { Config, ServerBundle } from './types';
 
-// SITECORE_JSS_APP_NAME env variable has been deprecated since v.21.6, SITECORE_SITE_NAME should be used instead
-const siteName = process.env.SITECORE_SITE_NAME || process.env.SITECORE_JSS_APP_NAME;
+const appName = process.env.SITECORE_JSS_APP_NAME || 'YOUR APP NAME';
+
+const siteName = process.env.SITECORE_SITE_NAME || appName;
 
 /**
  * The server.bundle.js file from your pre-built JSS app
  */
 
-const bundlePath = process.env.SITECORE_JSS_SERVER_BUNDLE || `../dist/${siteName}/server.bundle`;
+const bundlePath = process.env.SITECORE_JSS_SERVER_BUNDLE || `../dist/${appName}/server.bundle`;
 
 const serverBundle: ServerBundle = require(bundlePath);
 
@@ -28,8 +29,8 @@ export const config: Config = {
    */
   apiKey: process.env.SITECORE_API_KEY || serverBundle.apiKey || '{YOUR API KEY HERE}',
   /**
-   * The JSS application name defaults to providing part of the bundle path.
-   * If not passed as an environment variable or set here, any application name exported from the bundle will be used instead.
+   * The Sitecore site name.
+   * Required.
    */
   siteName: siteName || serverBundle.siteName,
   /**