+++ packages/create-sitecore-jss/src/templates/nextjs/src/lib/next-config/plugins/cors-header.js
@@ -9,7 +9,9 @@ const corsHeaderPlugin = (nextConfig = {}) => {
   }
   return Object.assign({}, nextConfig, {
     async headers() {
+      const extendHeaders = typeof nextConfig.headers === "function" ? await nextConfig.headers() : [];
       return [
+        ...await extendHeaders,
         {
           source: '/_next/:path*',
           headers: [{ key: 'Access-Control-Allow-Origin', value: config.sitecoreApiHost }],